+++ index.html
@@ -10,7 +10,7 @@
 </head>
 
 <body>
-  <script src="/js/functions.js"></script>
+
   <main>
 
     <!-- Фильтрация изображений от других пользователей -->
@@ -232,6 +232,6 @@
   <template id="messages">
     <div class="img-upload__message  img-upload__message--loading">Загружаем...</div>
   </template>
-
+  <script type="module" src="/js/main.js" ></script>
 </body>
 </html>
